requirement.txt
Django==2.2.2
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
django-rest-auth==0.9.5
